requirements.txt
python-telegram-bot[job-queue]==20.3
requests==2.31.0
Flask==2.3.3
python-dotenv==1.0.0
cachetools==5.3.1
psutil==5.9.5